requirements.txt
Django==4.0
gTTS==2.5.1
openai==1.32.0
pillow==10.3.0
tqdm==4.66.4

